最後のメンバーは脱退できないように変更

+++ app/controllers/memberships_controller.rb
@@ -6,6 +6,13 @@ class MembershipsController < ApplicationController
   end
 
   def destroy
+    group = current_user.groups.find(params[:group_id])
+    membership = current_user.memberships.find(params[:id])
+
+    if group.member_count == 1 && current_user.member_of?(group)
+      return redirect_to group_path(group), status: :see_other, alert: '最後のメンバーはグループを抜けられません。グループ削除を行なってください。'
+    end
+
     membership = current_user.memberships.find_by!(group_id: params[:group_id])
     membership.destroy!
